部分特征标签可以自动补充，移除手动定义相关代码；

+++ Samples/Dispatch Call/Sources/Component/MoveComponent.cs
@@ -31,6 +31,6 @@ namespace Game.Sample.DispatchCall
     [GameEngine.DeclareComponentClass("MoveComponent")]
     public class MoveComponent : GameEngine.CComponent, GameEngine.IUpdateActivation
     {
-        public float escape_time;
+        public int escape_time;
     }
 }
